+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 0.9.7 (xx-08-2026)
+- Added keyboard navigation in table paging inputs back (not sure when it disappeared).
 - Updated automatic world map naming to more likely pick up on real filenames.
 - Fixed automatic contenthash names sometimes naming empty BLPs incorrectly.